Try to fix redirect domain names in the deployment.
As documented in the README, it turns out that the App Service / App Gateway
setup create issues for the domain names in our redirects. Issuing redirects
is an important part of why we have this nginx server anyway, so I've set
something up that tries to Do The Right Thing. It's annoyingly complicated but
I can't devise anything cleaner.