Get /WebControl.aspx working again
Copied from legacy servers, with some minor futzing to retrieve assets
from wwtassets.org. I needed to configure nginx to return the right
content-type for this ASPX file -- there's a slim chance that maybe
we'll need to become cleverer if we ever get a different ASPX file that
wants to return a different content-type.